♻️ refactor(portal): drop the workspace HTML publish business slot
Keep packing and confirm in OSS. Publish UI becomes a no-op so cloud can
own the market-backed implementation without a business slot.
